requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
psycopg2==2.9.1
gunicorn==19.9.0
Pillow==9.1.0
Flask-Mail==0.9.1
